autopep8 for style
This changes the style to be in line with
PEP8 (https://www.python.org/dev/peps/pep-0008/), the official style
guide for Python code! It's by no means perscriptive or something
everyone follows to the letter, but it's useful to have something that
the community all sort of agrees on.

It's worth nothing that I didn't do this manually, I used a handy tool
called autopep8 that automatically reformats Python code to conform to
pep8! There are other tools, like pyflakes and pycodestyle, that you can
use to check that your code conforms as you write it.