use @property
The built in `property` decorator let's use functions as though they're
attributes! It's useful for situations like this where you want to just
think of the number of empty lines as a property of the board and not
have to worry about the fact that it's actually derived from an
array. You could probably do something simiar with other methods.